refactor: use prelude to restrict the visibility
